python/Mask_RCNN/requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow-gpu
keras
opencv-python
h5py
imgaug
IPython[all]